# api/openapi3.yaml
openapi: 3.0.0
info:
  contact:
    email: support@wishlist.example.com
    name: API Support
  description: >-
    A RESTful API for managing wish lists, gift items, and reservations.

    Features include user authentication, wish list management, gift item tracking, and reservation system.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Wish List API
  version: "1.1"
paths:
...
  /auth/login:
    post:
      description: Authenticate user with email and password
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/wish-list_internal_domain_user_delivery_http_dto.Lo\
                ginRequest"
        description: User login credentials
        required: true
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_user_delivery_http_dto.Au\
                  thResponse"
        "400":
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: User login
      tags:
        - Authentication
...
  /auth/register:
    post:
      description: Create a new user account with email and password
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/wish-list_internal_domain_user_delivery_http_dto.Re\
                gisterRequest"
        description: User registration information
        required: true
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_user_delivery_http_dto.Au\
                  thResponse"
        "400":
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "409":
          description: User with this email already exists
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      summary: Register a new user
      tags:
        - Authentication
...
  /protected/account:
    delete:
      description: Delete the authenticated user's account and all associated data.
        This action is irreversible.
      responses:
        "204":
          description: Account deleted successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Delete user account
      tags:
        - User
  /protected/export-data:
    get:
      description: Export the authenticated user's data in JSON format for compliance
        and personal records
      responses:
        "200":
          description: User data exported successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_user_delivery_http_dto.Ex\
                  portUserDataResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_user_delivery_http_dto.Er\
                  rorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_user_delivery_http_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Export user data
      tags:
        - User
  /protected/profile:
    get:
      description: Get the authenticated user's profile information
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_user_delivery_http_dto.Us\
                  erResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "404":
          description: User not found
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Get user profile
      tags:
        - User
    put:
      description: Update the authenticated user's profile information
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/wish-list_internal_domain_user_delivery_http_dto.Up\
                dateProfileRequest"
        description: Updated profile information
        required: true
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/wish-list_internal_domain_user_delivery_http_dto.Us\
                  erResponse"
        "400":
          description: Invalid request body or validation error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "404":
          description: User not found
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
      security:
        - BearerAuth: []
      summary: Update user profile
      tags:
        - User
...
servers:
  - url: //localhost:8080/api
components:
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...